pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "opsforge"
version = "1.0.0"
description = "A forge of essential tools for DevOps engineers and system administrators"
readme = "README.md"
authors = [
    {name = "Thomas Vincent", email = "info@thomasvincent.xyz"}
]
license = {file = "LICENSE"}
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Development Status :: 4 - Beta",
    "Intended Audience :: System Administrators",
    "Intended Audience :: Information Technology",
    "Topic :: System :: Monitoring",
    "Topic :: System :: Systems Administration",
    "Topic :: Utilities",
]
keywords = ["devops", "sysadmin", "monitoring", "automation", "dns", "http", "filesystem"]
dependencies = [
    "requests>=2.28.2",
    "click>=8.1.3",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "dnspython>=2.3.0",
    "paramiko>=3.0.0",
]
requires-python = ">=3.8"

[project.optional-dependencies]
dev = [
    "pytest>=7.3.1",
    "pytest-cov>=4.1.0",
    "black>=23.3.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.2.0",
    "pylint>=2.17.0",
    "pre-commit>=3.3.1",
]
docs = [
    "sphinx>=7.0.0",
    "sphinx-rtd-theme>=1.2.0",
]
gsheets = [
    "gspread>=5.9.0",
    "oauth2client>=4.1.3",
]
zenoss = [
    "zeep>=4.2.1",
]

[project.urls]
"Homepage" = "https://github.com/thomasvincent/opsforge"
"Bug Tracker" = "https://github.com/thomasvincent/opsforge/issues"
"Source Code" = "https://github.com/thomasvincent/opsforge"
"Documentation" = "https://opsforge.readthedocs.io"

[project.scripts]
opsforge = "opsforge.cli:main"
opsforge-http500 = "opsforge.http.http500:main"
opsforge-readonly = "opsforge.filesystem.readonly:main"
opsforge-dns = "opsforge.dns.manager:main"